Updated the data type of priority to u8 instead of u 32 (for storage efficiency)

+++ lib/libstd/Types.h
@@ -139,7 +139,7 @@ typedef unsigned short GroupID;
 /** Process Identification Number. */
 typedef u32 ProcessID;
 
-typedef u32 Priority;
+typedef u8 Priority;
 
 /**
  * Describes a device ID number.
